+++ HISTORY.rst
@@ -3,6 +3,17 @@
 Release History
 ---------------
 
+1.17.6 (2021-03-05)
++++++++++++++++++++
+
+**Improvements**
+
+- trade id added to context to prevent race condition and better visibility on live trades
+
+**Bug Fixes**
+
+- incorrect handling of replace on runner context fix (adds to live trade count)
+
 1.17.5 (2021-03-01)
 +++++++++++++++++++